test mar for speedmaps after folding into pipeline
