+++ rt_segment_speeds/scripts/Makefile
@@ -3,6 +3,7 @@ segmentize:
 	#pip install -r requirements.txt
 	python cut_stop_segments.py
 	python select_stop_segments.py
+	python cut_speedmap_segments.py    
 
 prep_roads:
 # run cut_road_segments and concatenate_road_segments once
@@ -21,9 +22,15 @@ rt_stop_times_pipeline:
 	python pipeline_rt_stop_times.py
 	python average_summary_speeds.py  
 
+speedmaps_pipeline:
+	python pipeline_speedmap.py
+	python average_speedmap_segment_speeds.py  
+    
 all_speeds_pipeline:
-	make speeds_pipeline && make rt_stop_times_pipeline
-
+	make speeds_pipeline
+	make rt_stop_times_pipeline
+	make speedmaps_pipeline
+    
 download_roads:
 	#pip install esridump
 	#esri2geojson https://geo.dot.gov/server/rest/services/Hosted/California_2018_PR/FeatureServer/0 ca_roads.geojson
