+++ rt_segment_speeds/segment_speed_utils/project_vars.py
@@ -16,8 +16,12 @@ analysis_date = rt_dates.DATES["apr2024"]
 oct2023_week = rt_dates.get_week("oct2023", exclude_wed=True)
 apr2023_week = rt_dates.get_week("apr2023", exclude_wed=True)
 apr2024_week = rt_dates.get_week("apr2024", exclude_wed=True)
-analysis_date_list  = apr2024_week
+
+weeks_available = [
+    rt_dates.apr2024_week, rt_dates.oct2023_week, rt_dates.apr2023_week
+]
+analysis_date_list  = [analysis_date]
 
 PROJECT_CRS = "EPSG:3310"
 ROAD_SEGMENT_METERS = 1_000
-SEGMENT_TYPES = ["stop_segments", "rt_stop_times"]
+SEGMENT_TYPES = ["stop_segments", "rt_stop_times", "speedmap_segments"]
